Material Management System/construction_erp/requirements.txt
asgiref==3.11.1
brotli==1.2.0
cffi==2.0.0
charset-normalizer==3.4.7
cssselect2==0.9.0
Django==6.0.3
django-cors-headers==4.9.0
django-filter==25.2
djangorestframework==3.17.1
djangorestframework_simplejwt==5.5.1
drf-yasg==1.21.15
et_xmlfile==2.0.0
fonttools==4.62.1
inflection==0.5.1
openpyxl==3.1.5
packaging==26.0
pillow==12.2.0
psycopg2-binary==2.9.11
pycparser==3.0
pydyf==0.12.1
PyJWT==2.12.1
pyphen==0.17.2
pytz==2026.1.post1
PyYAML==6.0.3
reportlab==4.4.10
sqlparse==0.5.5
tinycss2==1.5.1
tinyhtml5==2.1.0
tzdata==2026.1
uritemplate==4.2.0
weasyprint==68.1
webencodings==0.5.1
zopfli==0.4.1
